アニメ長パッチの型不一致でHarmony全体が止まっていたのを直し、カウント制限を復活させる。
GetCurrentAnimationLength は float。1パッチ失敗で後続が死なないよう型ごとに適用する。

+++ HypnosisCreatorCode/Patches/CombatFrameAnimPatch.cs
@@ -31,12 +31,13 @@ public static class CombatFrameAnimIdlePatch
 [HarmonyPatch(typeof(NCreature), nameof(NCreature.GetCurrentAnimationLength))]
 public static class CombatFrameAnimLengthPatch
 {
-    public static void Postfix(NCreature __instance, ref double __result)
+    // 本家は float (Single)。double だと PatchAll が例外で止まり後続パッチ全滅する。
+    public static void Postfix(NCreature __instance, ref float __result)
     {
         if (__result > 0) return;
         var len = CombatFrameAnimator.GetCurrentAnimLengthSeconds(__instance);
         if (len > 0)
-            __result = len;
+            __result = (float)len;
     }
 }
